requirements.txt
streamlit
langchain
pillow
pandas